pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rlqf"
version = "0.1.0"
description = "Reinforcement Learning with Quantum Feedback: actor-critic training of MM potentials with a QM critic"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1",
    "numpy>=1.24",
    "ase>=3.22",                  # Atomic Simulation Environment
    "mace-torch>=0.3",            # MACE interatomic potentials (actor)
    "orb-models>=0.4",            # OrbNet Denali / OrbMol (critic)
    "e3nn>=0.5",                  # Euclidean neural networks (equivariance)
    "torch-geometric>=2.4",       # Graph neural network primitives
    "pyyaml>=6.0",
    "tqdm>=4.65",
    "wandb>=0.16",                # Experiment tracking
    "hydra-core>=1.3",            # Configuration management
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.3",
    "mypy>=1.8",
]

[project.scripts]
rlqf-train = "rlqf.train:main"

[tool.setuptools.packages.find]
include = ["rlqf*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
